requirements.txt
streamlit
openai
pandas
matplotlib
openpyxl
python-dotenv
